pyproject.toml
[tool.poetry]
name = "antieroticspam"
version = "1.0.0"
description = "Telegram anti-erotic-spam bot"
authors = ["Team"]

[tool.poetry.dependencies]
# Only Python 3.12 is supported due to asyncpg C extensions compatibility
python = ">=3.12,<3.13"
python-telegram-bot = "^22.0"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
nudenet = "^3.0.0"
pillow = "^10.3.0"
opencv-python = "^4.10.0.82"
openai = "^1.25.0"
asyncpg = "^0.29.0"
python-multipart = "^0.0.9"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
prometheus-client = "^0.19.0"
pyjwt = "^2.8.0"
redis = "^5.0.1"
jinja2 = "^3.1.3"
pydantic = "^2.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"

[tool.poetry.scripts]
start = "bot.main:main"
test = "tests.run_tests:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"